requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4

# Validation & Forms
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Security
slowapi>=0.1.9
secure>=0.3.0

# External APIs
httpx>=0.26.0
stripe>=7.0.0

# Templates
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
